src/requirements.txt
bcrypt
email-validator
fastapi[all]==0.97.0
gunicorn
loguru
motor
PyJWT
pydantic
pydantic[email]
python-decouple
uvicorn[standard]